remove 'this area exactly' as a search option
Removing ‘this area exactly’ as a search option, because this implies
that areas have exact boundaries, when in reality they are often fuzzy
and hard to know precisely. Requiring a radius to be specified instead
removes this problem.

(Also removing the ‘Nationally’ option, as that is implied when the
location box is blank and the Distance option is disabled).