Convert dropdown to radio buttons on form
The main focus of this is to remove the downdowns (`<select>`) from the
Publish Vacancy form, as these aren’t ideal from an accessibility or
usability perspective (text can’t be resized as easily, and you can’t
see all the options without interacting with it). See guidance here:
https://paper.dropbox.com/doc/Select-boxes-v7PWDvPDV2tmc9pzDbm3s

In addition, “optional” fields should be avoided where possible, as
it’s easy to miss the “(optional)” parenthesis, and it’s preferable to
avoid asking people to have to select “not applicable” options.

With this in mind, I’ve made the following changes:

* Main subject become radio buttons instead of dropdown
* Pay sale becomes radio buttons which only appear if “Part of national
pay scale” ticked, rather than a dropdown.
* “This is a permanent position” checkbox moved underneath working
pattern.
* Leadership level dropdown becomes a set of radio buttons only visible
if “This is a leadership role is ticked”.