Feature: make the all the checkbox filters work
This implements all the checkbox filters, including salary range,
working pattern, education stage, leadership level, suitable for NQTs
and job share available.

The behaviour of these is generally an OR within the filter – e.g. if
you tick KS1 and KS2 you match jobs for either (or both) key stages,
and an AND between filters – e.g. if you tick KS1 and Part time then
vacancies have to match both.

The only exception is “Suitable for NQTs” and “Job share available”,
which act as independent binary choices (so if you tick both the
vacancies have to match both).

The keyword search works as before, and the location search doesn’t
work at all (yet).

As part of getting all this to work I ditched the ‘auto store’ data
function which saves all form data to the session as you fill out
forms, mainly because this kept confusing me and making debugging
harder. It’s a useful feature for initial prototyping, but I’m not sure
it’s needed any more at this stage?